1. 일부 리팩토링

+++ DefaultFrameWork_HSS/Assets/Scripts/Utillity/CSV/Editor/DataTableConvertWindow.cs
@@ -15,11 +15,17 @@ namespace HSS
         // ----- Param -----
         private readonly string readPath_All = "/DataTable/Excel";
         private readonly string savePath_DB = "/Resources/CSV/";
+        private readonly string SEARCH_DIRECTORY_PATH = "search_directory_path";
+
+        private string searchDirectory = "";
+        private string savedSearchDirectory = "";
+        private string projectDataTableDirectory = "Assets/DataTable/Excel";
+
         private static StringBuilder sbSuccess = new StringBuilder();
         private static StringBuilder sbFail = new StringBuilder();
-
         private static Regex regexComma = new Regex(",");
-        private readonly string SEARCH_DIRECTORY_PATH = "search_directory_path";
+
+        private StringBuilder sbProgress = new StringBuilder();
 
         // ----- Init -----
 
@@ -30,13 +36,6 @@ namespace HSS
             window.titleContent = new GUIContent("Convert All DataTable");
             window.Show();
         }
-
-        private string searchDirectory = "";
-        private string savedSearchDirectory = "";
-        private string projectDataTableDirectory = "Assets/DataTable/Excel";
-        private const string ERROR = "error";
-
-        private StringBuilder sbProgress = new StringBuilder();
 
         private void OnGUI()
         {
@@ -66,8 +65,6 @@ namespace HSS
 
         // ----- Get ----- 
 
-        private string GetDataPath_Folder(string folderName) => $"{projectDataTableDirectory}/{folderName}";
-
         private string GetDataPath_Application(string path) => $"{Application.dataPath}{path}";
 
         /// <summary>
@@ -93,10 +90,7 @@ namespace HSS
         {
             string name = string.Empty;
             string[] nameArray = filePath.Split(strSplit);
-            int maxLength = nameArray.Length;
-            if (maxLength > 0)
-                name = nameArray[maxLength - 1].Replace(strReplace, string.Empty);
-            return name;
+            return nameArray.Length > 0 ? nameArray[nameArray.Length - 1].Replace(strReplace, string.Empty) : name;
         }
 
         // ----- Main ----- 
@@ -111,35 +105,27 @@ namespace HSS
             DirectoryInfo di = new DirectoryInfo(GetDataPath_Application(readPath_All));
 
             // 루트 폴더안에 있는 엑셀 파일
-            string filePath;
-            foreach (System.IO.FileInfo file in di.GetFiles())
+            foreach (System.IO.FileInfo file in di.GetFiles("*.xls*"))
             {
-                if (file.Extension.ToLower().EndsWith(".xls") || file.Extension.ToLower().EndsWith(".xlsx"))
-                {
-                    filePath = GetDataPath_Application($"{readPath_All}/{file.Name}");
-                    ConvertToCSV(filePath, savePath_DB);
-                    HSSLog.Log("File.Name : " + file.Name);
-                }
+                ConvertToCSV(GetDataPath_Application($"{readPath_All}/{file.Name}"), savePath_DB);
+                HSSLog.Log("File.Name : " + file.Name);
             }
 
             // 루트 폴더내 다른 폴더에 있는 엑셀 파일
             foreach (DirectoryInfo folder in di.GetDirectories())
             {
                 HSSLog.Log("Folder.Name : " + folder.Name);
-                foreach (System.IO.FileInfo file in folder.GetFiles())
+                foreach (System.IO.FileInfo file in folder.GetFiles("*.xls*"))
                 {
-                    if (file.Extension.ToLower().EndsWith(".xls") || file.Extension.ToLower().EndsWith(".xlsx"))
-                    {
-                        filePath = GetDataPath_Application($"{readPath_All}/{folder.Name}/{file.Name}");
-                        ConvertToCSV(filePath, savePath_DB);
-                        HSSLog.Log("File.Name : " + file.Name);
-                    }
+                    ConvertToCSV(GetDataPath_Application($"{readPath_All}/{folder.Name}/{file.Name}"), savePath_DB);
+                    HSSLog.Log("File.Name : " + file.Name);
                 }
             }
 
             EditorUtility.DisplayDialog("Data Convert & Save Success.", sbSuccess.ToString(), "OK");
             if (sbFail.Length != 0)
                 EditorUtility.DisplayDialog("Data Convert Fail", sbFail.ToString(), "OK");
+
             AssetDatabase.Refresh();
         }
 
@@ -147,42 +133,37 @@ namespace HSS
         {
             sbProgress.Clear();
 
-            if (string.IsNullOrEmpty(projectDataTableDirectory)) return;
+            if (string.IsNullOrEmpty(projectDataTableDirectory)) 
+                return;
 
             DirectoryInfo dicInfo = new DirectoryInfo(searchDirectory);
-            List<System.IO.FileInfo> files = new List<System.IO.FileInfo>(dicInfo.GetFiles());
+            List<FileInfo> files = new List<FileInfo>(dicInfo.GetFiles());
 
-            foreach (System.IO.FileInfo info in files)
+            Dictionary<string, string> dicFileFolder = new Dictionary<string, string>
             {
-                if (info.Name.Contains("DS_Store")) continue;
+                { "ExampleFile" , "ExampleFolder"}
+            };
+
+            foreach (FileInfo info in files)
+            {
+                if (info.Name.Contains("DS_Store"))
+                    continue;
 
                 string name = GetFileName(info.Name, '.');
 
-                // 아래에 파일명과 폴더명을 추가함
-                string folderName = name switch
-                {
-                    "ExampleFile" => "ExampleFolder",
-                    _ => ERROR,
-                };
+                if (!dicFileFolder.TryGetValue(name, out string folderName))
+                    continue;
 
-                if (folderName.Equals(ERROR)) continue;
-
-                string destFolderPath = GetDataPath_Folder(folderName);
+                string destFolderPath = $"{projectDataTableDirectory}/{folderName}";
                 string destFilePath = $"{destFolderPath}/{info.Name}";
 
-#if UNITY_EDITOR_WIN
                 if (!Directory.Exists(destFolderPath))
                     AssetDatabase.CreateFolder(projectDataTableDirectory, folderName);
-#endif
+
                 DeleteFile($"{destFilePath}.xlsx");
                 DeleteFile($"{destFilePath}.meta");
 
-#if UNITY_EDITOR_WIN
                 File.Copy(info.FullName, destFilePath, true);
-#else
-                File.Copy(info.FullName, destFolderPath, true);
-#endif
-
                 sbProgress.AppendLine(info.Name);
             }
 
@@ -192,21 +173,17 @@ namespace HSS
 
         // ----- Logic ----- 
 
-        private void ConvertToCSV(string filePath, string savePath, int startRow = 3, Action callback = null)
+        // startRow 기본 3 -> 기획자의 임의 표기, 이름, 자료형
+        private void ConvertToCSV(string filePath, string savePath, int startRow = 3)
         {
-            // 기존 xlsx meta 파일 제거
-            //DeleteFile($"{filePath}.meta");
             HSSLog.Log($" From => {filePath} \n To => {savePath}");
 
             try
             {
                 using (var stream = File.Open(filePath, FileMode.Open, FileAccess.Read))
+                using (var reader = ExcelReaderFactory.CreateReader(stream))
                 {
-                    using (var reader = ExcelReaderFactory.CreateReader(stream))
-                    {
-                        ConvertProcess(reader.AsDataSet(), GetDataPath_Application(savePath), GetFileName(filePath, '/', ".xlsx"), startRow);
-                        callback?.Invoke();
-                    }
+                    ConvertProcess(reader.AsDataSet(), GetDataPath_Application(savePath), GetFileName(filePath, '/', ".xlsx"), startRow);
                 }
             }
             catch (Exception ex)
@@ -222,11 +199,12 @@ namespace HSS
                 int columns = result.Tables[0].Columns.Count;
                 int rows = result.Tables[0].Rows.Count;
                 StringBuilder sb = new StringBuilder();
+
                 for (int x = startRow; x < rows; x++)
                 {
                     for (int y = 0; y < columns; y++)
                     { 
-                        //? column
+                        // column
                         string str = result.Tables[0].Rows[x][y].ToString();
 
                         //// 비어있는 행 건너뛸 수 있음
@@ -239,7 +217,8 @@ namespace HSS
                             str = regexComma.Replace(str, "u002c");
 
                         if (str.StartsWith('{') && str.EndsWith('}'))
-                            str = string.Format("\"{0}\"", str);
+                            str = $"\"{str}\"";  //string.Format("\"{0}\"", str);
+
                         sb.Append(str);
                         if (y < columns - 1)
                             sb.Append(",");
@@ -260,7 +239,7 @@ namespace HSS
                 sbSuccess.Append(e.Message + "\n");
             }
         }
-    
+
         /// <summary>
         /// 파일 저장
         /// </summary>
@@ -274,11 +253,13 @@ namespace HSS
                 if (!Directory.Exists(filePath))
                     Directory.CreateDirectory(filePath);
 
-                string fileName = $"{name}.csv";
-                Stream fileStream = new FileStream(filePath + fileName, FileMode.OpenOrCreate, FileAccess.Write);
-                StreamWriter outStream = new StreamWriter(fileStream, new UTF8Encoding(true));
-                outStream.WriteLine(value);
-                outStream.Close();
+                File.WriteAllText($"{filePath}{name}.csv", value, Encoding.UTF8);
+
+                //string fileName = $"{name}.csv";
+                //Stream fileStream = new FileStream(filePath + fileName, FileMode.OpenOrCreate, FileAccess.Write);
+                //StreamWriter outStream = new StreamWriter(fileStream, new UTF8Encoding(true));
+                //outStream.WriteLine(value);
+                //outStream.Close();
             }
             catch (System.Exception e)
             {
@@ -297,10 +278,9 @@ namespace HSS
                 File.Delete(path);
         }
 
-        public void DrawSelectDirectoryUI(ref string path, string title, string saveKeyString = "", bool splitLayerGroup = true)
+        public void DrawSelectDirectoryUI(ref string path, string title)
         {
-            if (splitLayerGroup)
-                EditorGUILayout.BeginHorizontal();
+            EditorGUILayout.BeginHorizontal();
 
             path = EditorGUILayout.TextField(title, path);
             // 찾은 주소를 저장해서 불러오게 하려면 아래의 주석 해제
@@ -314,33 +294,13 @@ namespace HSS
             {
                 var selected = EditorUtility.OpenFolderPanel(title, path, "default");
                 if (string.IsNullOrEmpty(selected) != true)
-                {
                     path = selected;
-
-                    if (string.IsNullOrEmpty(saveKeyString) != true)
-                        EditorPrefs.SetString(saveKeyString, path);
-                }
             }
 
-            if (GUILayout.Button("Folder Open", GUILayout.Width(100)))
-            {
-                if (string.IsNullOrEmpty(path))
-                    EditorUtility.DisplayDialog("Error", "Empty Path", "OK");
-                else
-                {
-                    try
-                    {
-                        System.Diagnostics.Process.Start(path);
-                    }
-                    catch (System.Exception ex)
-                    {
-                        EditorUtility.DisplayDialog("Error", ex.Message, "OK");
-                    }
-                }
-            }
+            if (GUILayout.Button("Folder Open", GUILayout.Width(100)) && !string.IsNullOrEmpty(path))
+                System.Diagnostics.Process.Start(path);
 
-            if (splitLayerGroup)
-                EditorGUILayout.EndHorizontal();
+            EditorGUILayout.EndHorizontal();
         }
     }
 }
